renames uploadPhoto reducer to report
related #25
