+++ package.json
@@ -46,7 +46,6 @@
     "react-dom": "^15.4.2",
     "react-redux": "^5.0.3",
     "redux": "^3.6.0",
-    "redux-immutable": "^4.0.0",
     "redux-logger": "^2.8.2",
     "webpack": "^2.2.1"
   },
