+++ src/client/reducers/report.js
