+++ test/client/index.js
@@ -2,7 +2,7 @@
 require('./containers/upload_photo.js')
 
 // reducers
-require('./reducers/upload_photo.js')
+require('./reducers/report.js')
 
 // actions
 require('./actions/upload_photo.js')
