+++ test/client/reducers/report.js
@@ -1,22 +1,21 @@
 import test from 'tape'
 
-import reducer, { initialState }
-  from '../../../src/client/reducers/upload_photo.js'
+import reducer, { initialState } from '../../../src/client/reducers/report.js'
 import { SET_PHOTO } from '../../../src/client/action_types.js'
 
-test('uploadPhoto: initailState', t => {
+test('report: initailState', t => {
   t.equal(initialState.get('isPosting'), false, 'not posting request initially')
 
   t.end()
 })
 
-test('uploadPhoto reducer initialState', t => {
+test('report reducer initialState', t => {
   const newState = reducer(undefined, {})
-  t.equal(newState, initialState, 'uploadPhoto reducer')
+  t.equal(newState, initialState, 'report reducer')
   t.end()
 })
 
-test('uploadPhoto: SET_POSTING', t => {
+test('report: SET_POSTING', t => {
   const newState = reducer(initialState, { type: 'SET_POSTING' })
 
   t.equal(newState.get('isPosting'), true, 'sets isPosting to true')
@@ -24,7 +23,7 @@ test('uploadPhoto: SET_POSTING', t => {
   t.end()
 })
 
-test('uploadPhotoReducer: setPhoto', t => {
+test('report: SET_PHOTO', t => {
   const photoData = 'photo data'
   const newState = reducer(
     initialState,
